client/requirements.txt
requests==2.32.3
Pillow==10.3.0